For each book, create a deterministic, not reversible token for URLs.
We were having problems with non-ascii characters breaking:
https://github.com/NYPL-Simplified/web-reader-landing-page/issues/34

This creates a deterministic, but reversible (I don't think
being reversible is important) token based on author & title.
And uses for the book_desc directory.

It's not reversible because I remove non alphanumeric chars
from the base64 encoded string. I do that because
those chars may not be URL friendly.